+++ rest/src/main/java/net/boreeas/riotapi/rest/Incident.java
@@ -16,15 +16,18 @@
 
 package net.boreeas.riotapi.rest;
 
-import lombok.Getter;
+import lombok.Data;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * @author Malte Schütze
  */
-@Getter
-public class ParticipantPlayer {
-    private String matchHistoryUri;
-    private int profileIconId;
-    private String summonerName;
-    private long summonerId;
+@Data
+public class Incident {
+    private boolean incident;
+    private String created_at;
+    private long id;
+    private List<Message> messages = new ArrayList<>();
 }