+++ spectator/src/main/java/net/boreeas/riotapi/spectator/GamePool.java
@@ -39,7 +39,7 @@ public class GamePool {
 
         GameUpdateTask task = new GameUpdateTask(game, errorCallback);
         // Run once to pull all pending chunks, then run on repeat
-        task.run();
+        pool.schedule(() -> task.run(), 0, TimeUnit.MILLISECONDS);
         ScheduledFuture<?> self = pool.scheduleAtFixedRate(task, game.getLastChunkInfo().getNextAvailableChunk(), game.getChunkInterval(), TimeUnit.MILLISECONDS);
         task.setSelf(self);
 
@@ -50,10 +50,12 @@ public class GamePool {
         pool.shutdown();
     }
 
+    @Deprecated
     public static GamePool singleton(InProgressGame game) {
         return singleton(game, err -> {});
     }
 
+    @Deprecated
     public static GamePool singleton(InProgressGame game, Consumer<Exception> callback) {
         GamePool pool = new GamePool();
         pool.submit(game, callback);